down if isLocal

+++ database/migrations/2023_09_16_215201_create_products_table.php
@@ -31,6 +31,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('products');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('products');
+        }
     }
 };
