+++ database/migrations/2024_02_11_141455_create_groups_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('groups');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('groups');
+        }
     }
 };
