+++ database/migrations/2024_02_11_141845_add_collumn_group_id_to_product_table.php
@@ -21,9 +21,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropForeign(['group_id']);
-            $table->dropColumn(['group_id']);
-        });
+        if (app()->isLocal()) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropForeign(['group_id']);
+                $table->dropColumn(['group_id']);
+            });
+        }
     }
 };
