+++ database/migrations/2024_02_16_162825_create_sizes_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('sizes');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('sizes');
+        }
     }
 };
