+++ database/migrations/2024_02_16_164530_create_product_size_table.php
@@ -26,6 +26,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('product_size');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('product_size');
+        }
     }
 };
