+++ database/migrations/2024_02_23_154921_create_stickers_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('stickers');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('stickers');
+        }
     }
 };
